feat(course-items): Added CourseItems Controller/Service/Repository tests

+++ src/main/java/com/example/talktactics/repository/CourseItemRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface CourseItemRepository extends JpaRepository<CourseItem, Long> {
-    List<CourseItem> findByCourseId(int id);
+    List<CourseItem> findByCourseId(long id);
 }
